requirements.txt
discord.py
colorama
numpy
requests
aiohttp
bs4
pymongo
selenium
pyPrivnote
gtts
dnspython
image
utils
discord-webhook